setup.sh: Enhance the kernel driver selection when IOMMU is enabled
The PCI devices used for SPDK are bound with vfio-pci or
uio_pci_generic kernel drivers. In setup.sh, if the path /sys/kernel
/iommu_groups is not empty, vfio-pci kernel driver is the only choice;
otherwise uio_pci_generic is selected.

In system, IOMMU can be enabled but set to pass through. It means
IOMMU will not affect on the DMA transmission although IOMMU groups
has been configured. In this case both two kernel drivers are workable.

The patch will enhance the kernel drvier selection in SPDK. It allows
to bind PCI devices with vfio-pci or uio_pci_generic drivers although
IOMMU is enabled.

Change-Id: I540d8750bf837ce67b8bc8b516a1a3acb72c502c
Signed-off-by: tone.zhang <tone.zhang@arm.com>