crypto: ignore duplicate names in config or RPC
If a config file or RPC comes in with a bdev name or vbdev name
that has already been used, ignore that configuration attempt.
Alternatively we could fail the app as there is a config mistake
by the user but a message is printed.  Can change to failure if
others think that is a better course of action.

Addresses github #448

Change-Id: I62fe0f6afcd9859c49aacc33ccd72258ae16bbe3
Signed-off-by: Paul Luse <paul.e.luse@intel.com>