Update patch set 1
Patch Set 1: Code-Review-1

> Patch Set 1: Code-Review-1
> 
> (1 comment)

At this moment I found few issues with VPP/VCL 18.07.1 which makes it not working in our model and should be fixed in VPP like:

 1) Bind doesn't set local address properly (it's always set to INADDR_ANY),
 2) Accept doesn't accept new connections when server is started in non-blocking mode.

Both issues seems to be solved in a new VPP 18.10, however 18.10 is still under development and the newest stable release is 18.07.1 we decided to use. 

Thanks for a review, I think it's better to focus on Session API now and return to this patch when VPP 18.10 will be released.

Patch-set: 1
Reviewer: Tomasz Kulasek <1013008@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1