Update patch set 7
Patch Set 7: Code-Review-1

Seth, I'm lost here.   I did read your email, but I'm struggling to understand the result.   Producing the libspdk_env_dpk.so such that it contains the static DPDK libs breaks the ability to substitute another env implementation, and I think in your email you recognize that and also wrote "I think we have a path for doing that ahead of us."   Maybe that's where I'm a little confused.   By that, are you acknowledging that this patch does in fact break the capability of using an alternate env, but you have something else in mind forthcoming to restore that functionality?   If that's the case, will you be adding a new patch to this series so that it all goes in at once after they're all approved?

Patch-set: 7
Reviewer: Lance Hartmann <1011995@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1