Update patch set 8
Patch Set 8: Code-Review-1

Reiterating an exchange we had on IRC, there are 2 approaches to satisfying the ability for a consumer of the SPDK to craft and use their own alternate env.   In theory, one could keep the libspdk_env_spdk and implement their own rte_XXXX functions in place of the DPDK as long as functionally they performed the same and obviously adhered to the same API.  The second approach would be to replace the libspdk_env_spdk altogether thus conferring greater flexibility of the internal env implementation; i.e. it would be free from all DPDK constructs completely.

While the last patch removes libspdk_env_dpdk.so from the monolithic libspdk.so thus enabling the 2nd strategy, I cannot shake that I remain bothered that at this point we have a loss of symmetry between the static version of libspdk_env_dpdk and the shared version.   The static version (libspdk_env_dpdk.a) does not contain the DPDK objects, whereas the shared version (libspdk_env_dpdk.so) does.  I can't deny that troubles me.     Maybe it's just me, but I would assume that two libraries with the same name separated only by their suffix -- one static and the other shared -- would be functionally identical, but ours in this case are not.  I do feel we're going in the right direction here, but we're not quite there yet.  If someone wanted to link to our static spdk_env_dpdk then they'd have to write a make rule that also specified the DPDK libs.   If, however, they wanted to link to our shared spdk_env_dpdk, then they'd have to make a different rule because there'd be no need to specify the DPDK libs since in that case they are inside the shared version.   I think this loss of symmetry between the static and shared would be very confusing to the consumer of our libraries.

Patch-set: 8
Reviewer: Lance Hartmann <1011995@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1