Bdev/QoS: set the QoS thread to NULL when destroying
There is a race condition that the bdev is closed and the
related QoS cleanup has been queued as an event. The QoS
stuff on the bdev structure has been destroyed earlier
than this QoS cleanup event is executed. Later when cleaning
up the QoS, it refers to the bdev's QoS structure and hit
the NULL pointer.

Change-Id: Id797a0ced56ef2cebb93a71e0302006c12029814
Signed-off-by: GangCao <gang.cao@intel.com>