Update patch set 10
Patch Set 10:

Hi Darek,

Have you run any tests with this patch, to check total memory usage before and after this patch?  Just want to confirm that we're seeing minimal if any memory usage increase - or if there is a non-trivial increase that we understand it and are OK with it.

-Jim

Patch-set: 10
CC: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>