Update patch set 3
Patch Set 3:

I'm thinking about how to handle this case - where maybe 1 out of several controllers fail to initialize.

What if we changed the spdk_nvme_probe to always return 0.  But if we fail to attach to a controller for any reason, we still call the attach_cb with the trid, but pass NULL for the ctrlr!  Then the user knows which controllers failed.  This is just a thought - it is an implicit change in the API.

Patch-set: 3
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>