bdev/raid: do not set configuring raid bdev to offline
Previously, the difference between configuring and online/offline is
unclear, this patch just fixes it. The key difference should be the raid
bdev is whether registered. Online/offline is registered and can be seen
by upper layers, and configuring is the opposite.

According to the above, we should never set configuring raid bdev to
offline.

Change-Id: Id44ef6654e032993ffb8444e7e7ae3e43a9b0f16
Signed-off-by: wuzhouhui <wuzhouhui@kingsoft.com>