Update patch set 4
Patch Set 4: Code-Review-1

(5 comments)

Awesome! Some minor comments throughout, also I didn't test this.  It's easier just to let it run in CI and see how it works with the OS's that are on the 3 systems we have in CH that support crypto testing.  Then we can ask ITP to setup in Jenkins.  On the DPDK comment on line 6, we can do that after this is all working I think

Patch-set: 4
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1