Update patch set 1
Patch Set 1: Code-Review-1

I'm really excited about this patch. Thank you for working on this! I have several general questions before I really dive into the code review.

First, I think we need to figure out what happens when we don't have an available receive queue entry before we move forward with this. I re-read the relevant portions of the specification and this seems to be the key paragraph:

"The host RDMA Transport sets the value of RNR_RETRY based on the RDMA provider capability. If the RDMA provider supports RNR_RETRY, then the host should set RNR_RETRY to infinite (value of 7). If the RDMA provider does not support RNR_RETRY, then the host should set the RNR_RETRY to zero. The remote-peer shall match the host RNR_RETRY setting or fail the QP creation."

So if the host (initiator) supports RNR_RETRY, we should be requesting to turn it on. If it is enabled, then I think this shared receive queue strategy works out great. Does the Linux kernel host attempt to enable RNR_RETRY? The SPDK one doesn't, but we can change that, of course. We'll have to handle the case where the SPDK host attempts to enable RNR_RETRY, but the target doesn't support it and fails the QP creation by retrying without RNR_RETRY enabled. In the case where the host does not support RNR_RETRY, though, then I think we need to not use the shared receive queue for everything to work properly.

It's even trickier than the above makes it seem though, because we'll want to support the case where some hosts support RNR_RETRY and some don't, all at the same time. When a queue pair is first established, we need to check if RNR_RETRY is turned on. If it is, we need to set a flag in the spdk_nvmf_rdma_qpair that indicates it is using the shared receive queue. If it is not turned on, it needs to allocate sufficient resources within the queue pair itself. That really means that this should be a runtime check instead of a #define.

Second, our biggest problem area involves shutting down the target while I/O is outstanding. The current set of issues all involve finding completions on the shared completion queue after we've shut down a queue pair. We have lots of code to try and get everything to shut down cleanly, but there just doesn't appear to be any good way to guarantee that no more completions will show up on the completion queue for a given queue pair. I'm concerned that introducing a shared receive queue just doubles the problems in this area. Since you all are the RDMA experts, maybe you can enlighten us all on the right way to deal with this, for both the shared receive queue and the shared completion queue.

Thanks again!

Patch-set: 1
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1