Update patch set 1
Patch Set 1:

Regarding RNR_RETRY, I was thinking we should consider supporting existing initiators that do not set RNR_RETRY to 7. However, we might be early enough in the NVMe-oF deployment phase that we could get away with only supporting initiators that have RNR_RETRY set to 7. I'm mostly concerned with compatibility here - I want the target to work with a wide range of initiators (of varying quality). That's why I suggested doing a runtime check, but it would be a lot simpler if we checked for RNR_RETRY of 7 and if it isn't set, just deny the connection.

For the shutdown flow, you're right that we're not doing things quite like you outline. I have two additional questions that should make it clear enough for me to change it to the model you outline.

1) When setting a qp to state IBV_QPS_ERR, are we guaranteed to get a LAST_WQE_REACHED event? I thought I had seen cases where we don't get that event if the qp didn't have any outstanding operations when setting the state to IBV_QPS_ERR.

2) After the LAST_WQE_REACHED event arrives, if we get to a point where we poll the ibv_cq and it is entirely empty, does that guarantee that we won't get any additional completions regarding the qp that's been shut down? This is the part I was missing - I needed some way to know when no more completions would appear.

Thanks so much!

Patch-set: 1
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>