Update patch set 1
Patch Set 1:

We use this script to install dependencies onto systems, regardless of whether the user is using a VM. For example, I use this script to set up my bare metal Fedora machine.

There are at least three challenges here. One is that intel-ipsec-mb is not packaged but is a dependency of ours (when crypto is enabled). Another is that we need to install intel-ipsec-mb (which requires root) but we don't want to build it as root. We could build it as part of normal 'make' but we can't install it there because we aren't root. The third challenge is that we don't actually want to mess with this at all if the user doesn't plan to use crypto, but pkgdep.sh doesn't understand configuration options.

Patch-set: 1
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>