Update patch set 1
Patch Set 1:

Hi,

Regarding RNR_RETRY:
In scenarions "many to one" in case of SRQ, RNR is possible. Old clients with RNR_RETRY=0 can receive an error. If we enable SRQ by sdefault, that moves the issue into deployment area. Is it  commom , that SPDK based NVME-OF target serves different types of SPDK based initiators? If it''s, SRQ should be optional for a while. Currently, it's under conditinal compilation. We can change that to runtime configuration and set it to "disable" for a while by default.
As I see in Linux kernel SRQ usage is unconditional.


Shutdown flow:
I'll double check the behaviour of LAST_WQE_REACHED and return with answers. Of course, my answers will be relevant to Infiniband and RoCE. We don't have enough expertise with iWARP.
You mentioned there are some issues with current shutdown flow. Is it https://github.com/spdk/spdk/issues/298  ?
We can re-check this issue with this issue with and whithour SRQ in Infiniband/RoCE.

Thank you!

Patch-set: 1
Reviewer: Sasha Kotchubievsky <1014643@d5d70762-12d0-45a1-890d-524b12d3f735>