compress: Add vbdev module for compression
NOTE: Entire IO path under construction and incomplete but feel free
to review anything else :)

There is a mix of hack to test without the SPDK reduce library and
the start of integrating the reduce lib.

It will interface with a new library for metadata mmgmt and the DPDK
CompressDev API for compression/decompression and sit on top of any
valid vbdev/bdev.

Change-Id: I80332cddc09194b2e654bac14451133466d0fc26
Signed-off-by: Paul Luse <paul.e.luse@intel.com>