compress: Add vbdev module for compression
NOT READY FOR ANY KIND OF REVIEW. This is being based off the latest passthru template
and so far only has blind search and replace done.

It will interface with a new library for metadata mmgmt and the DPDK
CompressDev API for compression/decompression and sit on top of any
valid vbdev/bdev.

Change-Id: I80332cddc09194b2e654bac14451133466d0fc26
Signed-off-by: Paul Luse <paul.e.luse@intel.com>