Update patch set 9
Patch Set 9:

Can you also describe how this will be used?

For example, just getting the name of a controller may not be very useful.  For a namespace, users may need to know the controller it is associated with.  For a controller, they may need to know the transport ID.

Patch-set: 9
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>