Update patch set 38
Patch Set 38: Code-Review-1

(1 comment)

talked more with Ben about this - I think we should just require the user to specify the controller; the namespace ID is embedded in the SQE (Submission Queue Entry); this also makes it forward compatible with cases where we want to send a passthrough I/O command to a namespace that is *not* a bdev - for example, an OCSSD or zoned namespace.

Let's talk about this more at the dev summit.

Patch-set: 38
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1