Update patch set 1
Patch Set 1:

I'm not a huge fan of this as is. Some of the dependencies that are used in the build pool rely on being placed in sys_sgsw's home directory (see nvme-cli for example).
Of course this doesn't mean we can't just make sure we run it from that directory when setting up a new test machine. I guess I just don't understand the motivation for changing it.

Patch-set: 1
CC: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>