bdev/compress: Add configure option and build dependencies
Includes the required DPDK dependencies for SPDK block Reduce aka
Compression.

As ISA-L was added here initially for compression, it
has been included as a requirement in this patch for all x86 machines
so that it can be used by various utils across SPDK however it can
be disabled manually by using --without-isal. Note that this patch
supprots ISA-L for DPDK (reduce) only which requires an installation
of ISA-L.  Future work can modify the instalation requirement
accordingly (ie use libraries instead).

Configure also updated to check the various combinations of new options
that don't work together and fail when its not clear which option makes
sense, in the cases of missing deps we now just disable the option and
print a message..

Updated pkg_dep compression and crypto dependency check/install
to provide more user output.

Change-Id: Ic1ea3cbeb9373a7700f6f0c2a3194d65d6a34a41
Signed-off-by: paul luse <paul.e.luse@intel.com>