test/bdev: decrease malloc bdev size in bdev.conf.in
When trying to run the blockdev.sh test against a crypto bdev using qat,
I consistently got allocation failures on the malloc disks. I believe
that this is because the qat pmds require some amount of dma safe
memory. While the requirement is not huge, it is enough to fragment the
hugepages and make it impossible for us to allocate 7 contiguous 64MiB
buffers. The tests passed when this number was reduced to 32 MiB.

Change-Id: I66716a918123bc69e8b5d19412ac1f9c0b3dd462
Signed-off-by: Seth Howell <seth.howell@intel.com>
Reviewed-on: https://review.gerrithub.io/429534
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Chandler-Test-Pool: SPDK Automated Test System <sys_sgsw@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>