Update patch set 1
Patch Set 1:

That iWARP problem we've managed to get to the bottom of and it's an unrelated problem. Our issue on Mellanox cards is that I believe we occasionally see ibv_wq's come off the shared completion queue that do not contain the wr_id we initially provided. This only occurs when the ibv_wq has an error status.

I have a patch series out that attempts to catch this (and these checks really do hit!). Here's the one specifically for Mellanox cards, but the series contains another fix for Soft RoCE that is a similar but more severe problem. The series ends with a patch that switches the code over to use the shutdown sequence you prescribed (and all these bugs pop up when I switch it to that mechanism):

https://review.gerrithub.io/#/c/spdk/spdk/+/429275/

Patch-set: 1
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>