Update patch set 16
Patch Set 16:

Pawel, I should've explained my position better in the previous review.  My apologies.   I am advocating that we install the example binaries *only* as part of the base spdk package instead of producing an "spdk-examples" package at all.   That's because as you know we don't have the build-out-of-tree complexities worked out for the examples code, unless you've come up with some creative way around that.   This is in stark contrast to the dpdk-examples which does have that support.   I do see what you mean by the build temporary files (.o's, .d's, etc.), but that would be a moot point if we added an install target to each of the examples like is done for spdk_tgt.

Patch-set: 16
Reviewer: Lance Hartmann <1011995@d5d70762-12d0-45a1-890d-524b12d3f735>