Update patch set 4
Patch Set 4:

But this can already be determined from the other DEBUGLOG in this function.

Can you just modify the existing one to add either "enabled" or "disabled" based on the wce bit?  It seems better to print the value in both cases - not just when it's disabled.

Patch-set: 4
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>