Update patch set 16
Patch Set 16:

I am OK with not having an examples package for now, and installing *some* of the examples binaries with the main spdk package.

In my mind, this spec file is our first go at building an SPDK package.  As we work to get this upstreamed into Fedora, we can get their opinion on this.  I would be surprised if there aren't other changes we need to make to this spec file before it is accepted into Fedora.

Regarding which examples to integrate - here is my opinion:

nvme/perf
nvme/identify
nvme/hello_world
bdev/hello_world

Also test/bdev/bdevperf.

Which brings up the point that we probably need to do some reorganizing of some of these directories - things like bdevperf probably don't belong in the test directory anymore.  But that's for a different day.  :)

Patch-set: 16