Update patch set 1
Patch Set 1:

I'd especially like to get consensus on the name of this library.

I was originally calling it "sbzip" for "spdk block zip" but I received some concerns that this would indicate a new compress algorithm which this is not.  It is primarily an algorithm for saving the allocating blocks on disk and for laying out the metadata to translate between I/O requests to the logical volume and these allocated blocks on disk which have the compressed data.

"reduce" basically means "reduce the size of the data".  In the future, we could extend this to handle deduplication, although that is not touched on at all in the series of patches I will be posting.

We could also name the library "sdr" for "spdk data reducer" to shorten the name of the library.

Patch-set: 1
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>