Update patch set 10
Patch Set 10:

(5 comments)

Thanks guys for the detailed review, really appreciate it!  Ben, looking for your guidance on most of these as to whether or not the suggestions warrant a spin for this release tagged patch or if we should merge now and I can easily spin another without a tag on it (next week, I'm out the rest of this week) and baby sit it until it can merged post release. I'm totally fine either way, thanks!

Patch-set: 10
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>