Update patch set 8
Patch Set 8:

This also only does a linear search on the qpairs if the completion contains an error, so it has minimal impact. I would like to avoid it, but we'd have to consider dropping rxe as a test vehicle to make that happen.

Patch-set: 8
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>