Update patch set 8
Patch Set 8:

If we can deploy the fix for rxe to all of the test systems in the pool, we can drop this patch and the one after it. I don't think we should make an attempt to run against broken rxe versions.

Patch-set: 8
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>