Update patch set 1
Patch Set 1:

Hi Shuhei,

Sorry I did not quite understand your suggestion. The g_spdk_iscsi.MaxQueueDepth is from the global iSCSI setting with the default value of 64 where we currently hard coded. The TargetNode itself also has a queue depth which is for the session. The change here is that the session's queue depth is overrode with the global one.

For where the global MaxQueueDepth verification, it can be set like in our Conf file and if not just set with our default one.

Patch-set: 1
Reviewer: GangCao <1011207@d5d70762-12d0-45a1-890d-524b12d3f735>