Update patch set 2
Patch Set 2:

Just a note to say that I've tested this out and it doesn't solve the problem.

```
> spdk/scripts/vagrant/create_vbox.sh -s 8192 -d -p virtualbox fedora28

   default: Package git-2.17.2-1.fc28.x86_64 is already installed, skipping.
    default: Dependencies resolved.
    default: Nothing to do.
    default: Complete!
    default: mkdir: cannot create directory ‘spdk_repo/output’: Permission denied
The SSH command responded with a non-zero exit status. Vagrant
assumes that this means the command failed. The output for this command
should be in the log above. Please read the output to determine what
went wrong.

  SUCCESS!
```

Patch-set: 2